T: extract tests related to proc macro name resolution to separate class
